# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pykrx-mcp REST API",
    "description": "Korean stock market data API for ChatGPT Actions",
    "version": "0.5.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "description": "Health check endpoint.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/privacy-policy": {
      "get": {
        "summary": "Privacy Policy",
        "description": "Privacy policy page for ChatGPT Actions.",
        "operationId": "privacy_policy_privacy_policy_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/tools/get_stock_ohlcv": {
      "post": {
        "summary": "Get Stock Ohlcv",
        "description": "Get stock OHLCV data.",
        "operationId": "get_stock_ohlcv_tools_get_stock_ohlcv_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockOHLCVRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/get_market_ticker_list": {
      "post": {
        "summary": "Get Market Ticker List",
        "description": "Get list of stock tickers.",
        "operationId": "get_market_ticker_list_tools_get_market_ticker_list_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TickerListRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/get_market_ticker_name": {
      "post": {
        "summary": "Get Market Ticker Name",
        "description": "Get company name from ticker.",
        "operationId": "get_market_ticker_name_tools_get_market_ticker_name_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TickerNameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/get_market_cap_by_date": {
      "post": {
        "summary": "Get Market Cap By Date",
        "description": "Get market cap data.",
        "operationId": "get_market_cap_by_date_tools_get_market_cap_by_date_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketCapRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/get_market_fundamental_by_date": {
      "post": {
        "summary": "Get Market Fundamental By Date",
        "description": "Get fundamental data.",
        "operationId": "get_market_fundamental_by_date_tools_get_market_fundamental_by_date_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FundamentalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/get_market_trading_value_by_date": {
      "post": {
        "summary": "Get Market Trading Value By Date",
        "description": "Get investor trading value (supply/demand analysis).",
        "operationId": "get_market_trading_value_by_date_tools_get_market_trading_value_by_date_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TradingValueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/get_etf_ohlcv_by_date": {
      "post": {
        "summary": "Get Etf Ohlcv By Date",
        "description": "Get ETF OHLCV data.",
        "operationId": "get_etf_ohlcv_by_date_tools_get_etf_ohlcv_by_date_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ETFOHLCVRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/get_etf_ticker_list": {
      "post": {
        "summary": "Get Etf Ticker List",
        "description": "Get list of ETF tickers.",
        "operationId": "get_etf_ticker_list_tools_get_etf_ticker_list_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ETFTickerListRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ETFOHLCVRequest": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "start_date": {
            "type": "string",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "title": "End Date"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "start_date",
          "end_date"
        ],
        "title": "ETFOHLCVRequest"
      },
      "ETFTickerListRequest": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          }
        },
        "type": "object",
        "required": [
          "date"
        ],
        "title": "ETFTickerListRequest"
      },
      "FundamentalRequest": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "start_date": {
            "type": "string",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "title": "End Date"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "start_date",
          "end_date"
        ],
        "title": "FundamentalRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MarketCapRequest": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "start_date": {
            "type": "string",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "title": "End Date"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "start_date",
          "end_date"
        ],
        "title": "MarketCapRequest"
      },
      "StockOHLCVRequest": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "start_date": {
            "type": "string",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "title": "End Date"
          },
          "adjusted": {
            "type": "boolean",
            "title": "Adjusted",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "start_date",
          "end_date"
        ],
        "title": "StockOHLCVRequest"
      },
      "TickerListRequest": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "market": {
            "type": "string",
            "title": "Market",
            "default": "KOSPI"
          }
        },
        "type": "object",
        "required": [
          "date"
        ],
        "title": "TickerListRequest"
      },
      "TickerNameRequest": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          }
        },
        "type": "object",
        "required": [
          "ticker"
        ],
        "title": "TickerNameRequest"
      },
      "TradingValueRequest": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "start_date": {
            "type": "string",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "title": "End Date"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "start_date",
          "end_date"
        ],
        "title": "TradingValueRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}